feat: split commands into self containing files
fix: resolve issue with template path
