+++ package.json
@@ -52,6 +52,7 @@
     "execa": "^6.1.0",
     "listr": "^0.14.3",
     "lodash.merge": "^4.6.2",
+    "random-words": "^1.3.0",
     "schema-utils": "^4.0.0",
     "which": "^2.0.2",
     "write-json-file": "^5.0.0"
