+++ src/cli/options.js
@@ -2,9 +2,11 @@ import { Option } from 'commander';
 
 const defaultConfigFile = 'wpnd.config.json';
 
-const programConfigFile = new Option(
+export const programConfigFile = new Option(
   '-c,--config <file>',
   'path to config file'
 ).default(defaultConfigFile);
 
-export default programConfigFile;
+export const showLogs = new Option('-v, --verbose', 'show docker logs').default(
+  false
+);
